1/6/2023

+++ University Course And Result Management System/server/UniversityCourseAndResultManagementSystem.Model/Teacher.cs
@@ -11,8 +11,8 @@
         public Guid DesignationId { get; set; }
         public Department Department { get; set; }
         public Guid DepartmentId { get; set; }
-        public int CreditToBeTaken { get; set; }
-        public int CreditTaken { get; set; }
+        public float CreditToBeTaken { get; set; }
+        public float CreditTaken { get; set; }
         public List<AssignedCourse> AssignedCourses { get; set; }
     }
 }
